requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1
pymongo==4.8.0

timm==1.0.9
facenet-pytorch==2.6.0
opencv-python==4.10.0.84
ffmpeg-python==0.2.0
tqdm==4.66.5
Pillow==10.2.0
numpy==1.26.4
torch
torchvision
